@mapKey

+++ src/main/java/com/kfit/core/repository/DepartmentDao.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by davi on 2017/7/16.
  */
-public interface DepartmentDao extends JpaRepository<Department,Integer> {
+public interface DepartmentDao extends JpaRepository<Department,Long> {
 }
